PillowWriter fix
Now it works on both Colab and Local machine as it is using `PillowWriter` instead of `ffmpeg`. The code execution is tiny bit slower and it creates and saves the thing to `temp.gif` file as I could not find how to make temporary file working without permission error on Windows.